+++ services/ui-src/src/index.tsx
@@ -1,5 +1,5 @@
 import ReactDOM from "react-dom";
-import "index.scss";
+import "./styles/index.scss";
 import App from "App";
 import * as serviceWorker from "serviceWorker";
 import { BrowserRouter as Router } from "react-router-dom";
@@ -8,7 +8,7 @@ import { QueryProvider } from "query";
 import { ReactQueryDevtools } from "react-query/devtools";
 import config from "config";
 import { ChakraProvider } from "@chakra-ui/react";
-import { theme } from "theme";
+import { theme } from "styles/theme";
 import { UserProvider, ApiProvider } from "hooks/authHooks";
 
 Amplify.configure({
